+++ legacy/modules/legacy_info.py
@@ -7,17 +7,13 @@
 import git
 from legacytl.tl.types import Message
 from legacytl.utils import get_display_name
-import requests
-import os
 from .. import loader, utils, version
-from ..inline.types import InlineQuery
 import platform as lib_platform
 import getpass
 
+
 @loader.tds
 class LegacyInfoMod(loader.Module):
-    """Show userbot info"""
-
     strings = {"name": "LegacyInfo"}
 
     def __init__(self):
@@ -26,7 +22,6 @@ class LegacyInfoMod(loader.Module):
                 "custom_message",
                 doc=lambda: self.strings("_cfg_cst_msg"),
             ),
-
             loader.ConfigValue(
                 "banner_url",
                 "https://i.postimg.cc/NfKrrv54/41-2807-ED0.gif",
@@ -49,7 +44,7 @@ class LegacyInfoMod(loader.Module):
             utils.escape_html(get_display_name(self._client.legacy_me)),
         )
         build = utils.get_commit_url()
-        _version = f'<i>{version.__version__}</i>'
+        _version = f"<i>{version.__version__}</i>"
         prefix = f"«<code>{utils.escape_html(self.get_prefix())}</code>»"
 
         platform = utils.get_named_platform()
@@ -63,14 +58,16 @@ class LegacyInfoMod(loader.Module):
             ("🕶", "<emoji document_id=5407025283456835913>📱</emoji>"),
             ("💎", "<emoji document_id=5471952986970267163>💎</emoji>"),
             ("🛡", "<emoji document_id=5422712776059534305>🌩</emoji>"),
+            ("🦈", "<emoji document_id=5201956344717936264>🦈</emoji>"),
             ("🌼", "<emoji document_id=5224219153077914783>❤️</emoji>"),
             ("🎡", "<emoji document_id=5226711870492126219>🎡</emoji>"),
             ("🐧", "<emoji document_id=5361541227604878624>🐧</emoji>"),
             ("🦊", "<emoji document_id=5283051451889756068>🦊</emoji>"),
-            ("🧨", "<emoji document_id=5379774338733994368>🧨</emoji>")
+            ("🧨", "<emoji document_id=5379774338733994368>🧨</emoji>"),
         ]:
             platform = platform.replace(emoji, icon)
-        return (self.config["custom_message"].format(
+        return (
+            self.config["custom_message"].format(
                 me=me,
                 version=_version,
                 build=build,
@@ -84,19 +81,20 @@ class LegacyInfoMod(loader.Module):
                 hostname=lib_platform.node(),
                 user=getpass.getuser(),
                 kernel=lib_platform.uname().release,
-                os=lib_platform.system()
-            ) if self.config["custom_message"] and "-d" not in args
+                os=lib_platform.system(),
+            )
+            if self.config["custom_message"] and "-d" not in args
             else (
-                f'<b>{{}}</b>\n\n<b>{{}} {self.strings("owner")}:</b> {me}\n\n<b>{{}}'
-                f' {self.strings("version")}:</b> {_version} {build}\n<b>{{}}'
-                f' {self.strings("branch")}:'
+                f"<b>{{}}</b>\n\n<b>{{}} {self.strings('owner')}:</b> {me}\n\n<b>{{}}"
+                f" {self.strings('version')}:</b> {_version} {build}\n<b>{{}}"
+                f" {self.strings('branch')}:"
                 f"</b> <code>{version.branch}</code>\n{upd}\n\n<b>{{}}"
-                f' {self.strings("prefix")}:</b> {prefix}\n<b>{{}}'
-                f' {self.strings("uptime")}:'
+                f" {self.strings('prefix')}:</b> {prefix}\n<b>{{}}"
+                f" {self.strings('uptime')}:"
                 f"</b> {utils.formatted_uptime()}\n\n<b>{{}}"
-                f' {self.strings("cpu_usage")}:'
+                f" {self.strings('cpu_usage')}:"
                 f"</b> <i>~{await utils.get_cpu_usage_async()} %</i>\n<b>{{}}"
-                f' {self.strings("ram_usage")}:'
+                f" {self.strings('ram_usage')}:"
                 f"</b> <i>~{utils.get_ram_usage()} MB</i>\n<b>{{}}</b>"
             ).format(
                 *map(
@@ -130,9 +128,7 @@ class LegacyInfoMod(loader.Module):
                 await self._render_info(False, args),
             )
         else:
-            await utils.answer(
-                message, await self._render_info(False, args)
-            )
+            await utils.answer(message, await self._render_info(False, args))
 
     @loader.command()
     async def ubinfo(self, message: Message):
@@ -145,4 +141,3 @@ class LegacyInfoMod(loader.Module):
 
         self.config["custom_message"] = args
         await utils.answer(message, self.strings("setinfo_success"))
-